src/nn/layers.py
import cupy as cp
import numpy as np

class EncryptedFeatureExtractor:
    def __init__(self, rns_context, encoder):
        self.rns = rns_context
        self.encoder = encoder
        
        # Simulated pre-trained ML weights for the 1024 drone sensors
        # e.g., alternating weights to detect specific anomalies
        self.weights = np.array([(i % 5) + 1 for i in range(encoder.N)], dtype=np.uint32)
        self.pt_weights = self.encoder.encode(self.weights)

    def forward(self, ct):
        print("[Feature Extractor] Applying pre-trained ML weights to encrypted SIMD array...")
        # Y = W * X (Pointwise SIMD Multiplication across 5 RNS Limbs)
        res0, res1 = self.rns.he_mul_plain(ct, self.pt_weights)
        return res0, res1
